Move restate-service-protocol types to restate-types
Moving the restate-service-protocol types to restate-types makes
it simpler to share them across different crates.
