+++ crates/service-protocol/Cargo.toml
@@ -11,11 +11,10 @@ publish = false
 default = []
 
 awakeable-id = ["dep:base64", "dep:restate-base64-util", "dep:restate-types"]
-codec = ["protocol", "dep:restate-types", "dep:paste"]
+codec = ["dep:restate-types", "dep:paste"]
 discovery = ["dep:serde", "dep:serde_json", "dep:regress", "dep:tracing", "dep:codederror", "dep:restate-errors", "dep:restate-schema-api", "dep:hyper", "dep:restate-service-client", "dep:restate-types", "dep:tokio"]
-message = ["protocol", "dep:restate-types", "dep:bytes-utils", "dep:codederror", "dep:restate-errors", "dep:size", "dep:tracing"]
+message = ["dep:restate-types", "dep:bytes-utils", "dep:codederror", "dep:restate-errors", "dep:size", "dep:tracing"]
 mocks = ["awakeable-id"]
-protocol = []
 
 [dependencies]
 restate-base64-util = { workspace = true, optional = true }
@@ -48,12 +47,3 @@ test-log = { workspace = true }
 tokio = { workspace = true }
 tracing-subscriber = { workspace = true }
 uuid = { workspace = true }
-
-[build-dependencies]
-prost-build = { workspace = true }
-prettyplease = "0.2"
-schemars = { workspace = true }
-serde_json = { workspace = true }
-syn = "2.0"
-typify = { version = "0.0.16" }
-jsonptr = "0.4.7"
