--- crates/service-protocol/README.md
@@ -1,7 +1,0 @@
-# Service protocol implementation
-
-To update the subtree, from the root directory of the project:
-
-```shell
-git subtree pull --prefix crates/service-protocol/service-protocol git@github.com:restatedev/service-protocol.git main --squash
-```
