--- crates/service-protocol/service-protocol/.gitignore
@@ -1,2 +1,0 @@
-.idea
-*.iml
