--- crates/service-protocol/service-protocol/.prettierrc.toml
@@ -1,3 +1,0 @@
-embeddedLanguageFormatting = "off"
-proseWrap = "always"
-printWidth = 120
