--- crates/service-protocol/service-protocol/.protolint.yaml
@@ -1,12 +1,0 @@
-lint:
-  rules:
-    remove:
-      - ENUM_FIELD_NAMES_PREFIX
-
-  rules_option:
-    max_line_length:
-      max_chars: 180
-      tab_chars: 2
-
-    indent:
-      style: 2
