--- crates/service-protocol/service-protocol/README.md
@@ -1,13 +1,0 @@
-# Restate Service Protocol
-
-This repo contains specification documents and Protobuf schemas of the Restate Service Protocol.
-
-* [Service invocation protocol specification](./service-invocation-protocol.md)
-
-## Development
-
-To format the spec document:
-
-```shell
-npx prettier -w service-invocation-protocol.md
-```
