+++ crates/types/README.md
@@ -1,0 +1,7 @@
+# Service protocol types
+
+To update the subtree, from the root directory of the project:
+
+```shell
+git subtree pull --prefix crates/types/service-protocol git@github.com:restatedev/service-protocol.git main --squash
+```
