+++ crates/types/src/lib.rs
@@ -20,6 +20,7 @@ pub mod arc_util;
 pub mod art;
 pub mod config;
 pub mod deployment;
+pub mod endpoint_manifest;
 pub mod epoch;
 pub mod errors;
 pub mod identifiers;
@@ -33,6 +34,8 @@ pub mod net;
 pub mod nodes_config;
 pub mod partition_table;
 pub mod retries;
+pub mod service_discovery;
+pub mod service_protocol;
 pub mod state_mut;
 pub mod storage;
 pub mod subscription;
