+++ crates/types/src/service_discovery.rs
@@ -1,0 +1,14 @@
+// Copyright (c) 2024 - Restate Software, Inc., Restate GmbH.
+// All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+include!(concat!(
+    env!("OUT_DIR"),
+    "/dev.restate.service.discovery.rs"
+));
